+++ App/Componentes/seccionCartelera.mjs
@@ -1,0 +1,14 @@
+const cateleraApp = () => {
+    const $cartelera = document.getElementById("id-cartelera");
+    const $secInicio = document.getElementById("sec-inicio");
+    const $secCartelera = document.getElementById("sec-cartelera");
+    const $secCombo = document.getElementById("sec-combo");
+    
+    $cartelera.addEventListener("click", () => {
+        $secInicio.style.display = "none";
+        $secCartelera.style.display = "block";
+        $secCombo.style.display = "none";
+    });
+
+}
+export default cateleraApp;
